[Refactor] Move package to storage
