+++ android/app/src/main/java/com/phicdy/totoanticipation/model/storage/GameListStorageImpl.java
@@ -1,4 +1,4 @@
-package com.phicdy.totoanticipation.model;
+package com.phicdy.totoanticipation.model.storage;
 
 import android.content.Context;
 import android.content.SharedPreferences;
@@ -6,6 +6,7 @@ import android.support.annotation.NonNull;
 
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
+import com.phicdy.totoanticipation.model.Game;
 
 import java.util.ArrayList;
 import java.util.List;
